fixed broken tests in BibliotecaAppTest
The tests got buggy after doing a dress rehearsal from command line so needed fixing.
One important change to fix stuff included overloading constructor of BibliotecaApp which was necessary because we want to sometimes inject the mocked menu for testing but also not need to.
So now an arity of 1 or 0 is available to run the app.